refactor(api): extract lifecycle components into multiple files
